Login error and JSON fixed (It took me 3 hours, I just want to die)

+++ ClientLibrary/Helper/ITokenService.cs
@@ -5,7 +5,7 @@
         Task<string> GetJWTTokenAsync(string key);
         Task<string> GetRefreshTokenAsync(string key);
         string FormToken(string jwt, string refresh);
-        Task SetCookie(string key, string value, int days, string path);
+        Task SetCookie(string key, string value, long days, string path);
         Task RemoveCookie(string key);
     }
 }
